requirements.txt
bottle==0.11.4
distribute==0.6.31
httplib2==0.7.7
python-instagram==0.8.0
simplejson==2.6.2
wsgiref==0.1.2
